Where do I begin? Include all the static assets
Sorry.

I'm really very sorry.

Static has had styles removed, which has broken the Transactions
Explorer. Our options are:

1) Start using govuk_template

or

2) This

Integrating govuk_template with this repo would be a hell of a lot
of work, hence this commit.

Include all the static assets pinned to the specific version that
is currently in production on the Transactions Explorer.